Extract shared category form utilities
